requirements.txt
agent-starter-pack==0.1.7
annotated-types==0.7.0
anyio==4.9.0
arrow==1.3.0
authlib==1.6.0
binaryornot==0.4.4
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
cloudpickle==3.1.1
cookiecutter==2.5.0
cryptography==45.0.4
docstring-parser==0.16
exceptiongroup==1.3.0
fastapi==0.115.13
google-adk==1.5.0
google-api-core==2.24.2
google-api-python-client==2.174.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-cloud-aiplatform==1.99.0
google-cloud-appengine-logging==1.6.2
google-cloud-audit-log==0.3.2
google-cloud-bigquery==3.30.0
google-cloud-core==2.4.3
google-cloud-logging==3.12.1
google-cloud-resource-manager==1.14.2
google-cloud-secret-manager==2.24.0
google-cloud-speech==2.33.0
google-cloud-storage==2.19.0
google-cloud-trace==1.16.2
google-crc32c==1.7.0
google-genai==1.22.0
google-resumable-media==2.7.2
googleapis-common-protos==1.69.2
graphviz==0.21
greenlet==3.2.3
grpc-google-iam-v1==0.14.2
grpcio==1.71.0
grpcio-status==1.71.0
h11==0.16.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
httpx-sse==0.4.1
idna==3.10
importlib-metadata==8.7.0
jinja2==3.1.6
markdown-it-py==3.0.0
markupsafe==3.0.2
mcp==1.9.4
mdurl==0.1.2
numpy==2.2.4
opentelemetry-api==1.34.1
opentelemetry-exporter-gcp-trace==1.9.0
opentelemetry-resourcedetector-gcp==1.9.0a0
opentelemetry-sdk==1.34.1
opentelemetry-semantic-conventions==0.55b1
packaging==24.2
proto-plus==1.26.1
protobuf==5.29.4
pyasn1==0.6.1
pyasn1-modules==0.4.1
pycparser==2.22
pydantic==2.11.7
pydantic-core==2.33.2
pydantic-settings==2.10.1
pygments==2.19.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
python-slugify==8.0.4
pyyaml==6.0.2
requests==2.32.4
rich==13.7.1
rsa==4.9
shapely==2.0.7
six==1.17.0
sniffio==1.3.1
sqlalchemy==2.0.41
sse-starlette==2.3.6
starlette==0.46.2
tenacity==8.5.0
text-unidecode==1.3
types-python-dateutil==2.9.0.20241206
typing-extensions==4.12.2
typing-inspection==0.4.1
tzlocal==5.3.1
uritemplate==4.2.0
urllib3==2.3.0
uvicorn==0.34.3
websockets==15.0.1
zipp==3.23.0
